Move SourceMetadata into the storage package
