+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/vhs/SourceMetadata.scala
@@ -1,3 +1,3 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.storage.vhs
 
 case class SourceMetadata(sourceName: String)
